Add timer into activity

+++ app/src/main/java/org/pixelexperience/recorder/sounds/OnTimerUpdatedListener.java
@@ -16,5 +16,5 @@
 package org.pixelexperience.recorder.sounds;
 
 interface OnTimerUpdatedListener {
-    void onTimerUpdated(int seconds);
+    void onTimerUpdated(long seconds);
 }
